chore: refactor some OCI-SIF creation code
A lot of code that is concerned with writing an OCI image to disk as
OCI-SIF was located inside `internal/pkg/client/ocisif`. This isn't a
great place as we expand to e.g. writing OCI-SIF data containers and
later manipulating oci-sif overlays.

Perform a very coarse initial refactoring, bringing the code to write
OCI-SIF images into `pkg/internal/ocisif` alongside the recent data
container code.

Yes... the function names are a bit horrible. `WriteXXX` will be
refactored to be a single function taking options, rather than several
related functions.

We are also in a better position to add more test code around the
OCI-SIF creation parts now.
