+++ internal/pkg/ocisif/testdata/Test_fileTARWriter/FileInRootDir.golden
