+++ internal/pkg/ocisif/testdata/Test_fileTARWriter/FileInSubDir.golden
