+++ internal/pkg/ocisif/testdata/Test_fsTARWriter/RootDir.golden
