+++ internal/pkg/ocisif/testdata/Test_fsTARWriter/SubDir.golden
