+++ internal/pkg/ocisif/testdata/Test_newDataContainerFromFSPath/FileInRootDir.golden
