+++ internal/pkg/ocisif/testdata/Test_newDataContainerFromFSPath/FileInSubDir.golden
