+++ internal/pkg/ocisif/testdata/Test_newDataContainerFromFSPath/RootDir.golden
