+++ internal/pkg/ocisif/testdata/Test_newDataContainerFromFSPath/SubDir.golden
